Create the apical JSON file at the end of synthesis_morphologies
The JSON file that contains the apical section ID for all morphologies
is now created in the 'finalize()' of synthesis_morphologies.
The app introduces a new argparse option '--out-apical' that specifies the path
to the JSON file to be created.
See https://bbpteam.epfl.ch/project/issues/browse/BBPP82-13 for the context
of this task

Also:
* stop using 'six' since the package is now python 3 only
* Unpin mpi4py

Change-Id: I7dcd296f76dbf2e16501bb3f60ab0f2ac60a51ec