Fix bug in writing the apical JSON file
The wrong section ID was being written in the file as what matters
is not the section ID but the section ID relative to the first apical
section.

https://bbpteam.epfl.ch/project/issues/browse/BBPP82-13?focusedCommentId=110967&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-110967

Also synthesis in H5 format has been disabled since we can't get the
NEURON apical section IDs for H5 files (since they can't be read by NEURON).

H5 files will be re-enabled in the next commit.

Change-Id: Ifc00b4fc8dfac895d0b82c68d2e28f3288959833