Writing to H5 must be accompagnied by another file format
as we can't use NEURON with H5 files so we can't get the NEURON section ID.
This is circumvented by getting the section ID from the file written in
another file format. This works because morphio sanitization before writing
the file ensures the section ID equivalence.

This commits also perform a cleanup of the synthesize_morphologies file.
The biggest change is that since TNS synthesis can fail, 10 attempts are
allowed before giving up and skipping it.

Change-Id: Id9cbe1ee55507fc22a83956470b3651e4a10c6bc