Export scores choose_morphologies, add gaussian bias and add jitter
Scores export:
These files are usefull to understand why a given morphology
was chosen or why no morphology could be chosen.

Gaussian bias:
This new option intends to penalize axons more strongly when the scale factor
is far from 1.

Jitter:
Instead of using the clones, it is now possible to apply jitter directly
during the synthesis process applied to the repaired cells.

Change-Id: I15a7bb2dbf7d2449a557a94e0473e3173ef088e0